Time: 0 ms (100.00%) | Memory: 7.2 MB (23.93%) - LeetSync

+++ 476-number-complement/number-complement.cpp
@@ -2,8 +2,8 @@ class Solution {
 public:
     int findComplement(int num) {
         long long ans = 0;
-        long long binary[64];
-        long long i = 0;
+        int binary[64];
+        int i = 0;
         while(num > 0){
             binary[i] = num % 2;
             if(binary[i] == 1)
@@ -14,7 +14,7 @@ public:
             num /= 2;
             i++;
         }
-        long long j = 0;
+        int j = 0;
         long long factor = 1;
         while(j < i){
             if(binary[j] == 1)
